requirements.txt
# Modern Treasury Dataloader — Direct Dependencies
# Pinned to exact versions for reproducible installs.
# starlette is provided transitively by fastapi.

modern-treasury==1.72.0
fastapi==0.135.1
uvicorn[standard]==0.42.0
pydantic==2.12.4
pydantic-settings==2.12.0
pyyaml==6.0.3
sse-starlette==3.3.3
tenacity==9.1.4
loguru==0.7.3
jinja2==3.1.6
python-multipart==0.0.22
httpx>=0.27.0
faker>=40.0
pyngrok>=7.5.1
sqlalchemy[asyncio]==2.0.48
aiosqlite==0.20.0
alembic==1.14.0
pytest-asyncio==0.25.0
